requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.39.1
dnspython==2.6.1
email_validator==2.1.1
Flask==2.2.5
flask-mongoengine==1.0.0
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.28.2
pymongo==4.6.2
python-dotenv==1.0.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
